+++ coo/utils.py
@@ -86,5 +86,5 @@ def tweet_template(update: str, template: str) -> str:
 
 def check_schedule_len(schedule: tuple):
     """Raises a ScheduleError if the len of the tuple is less tan 3."""
-    if len(schedule) != 3:
+    if len(schedule) is not 3:
         raise ScheduleError(ScheduleError.tupleLenError)